Move dispatch to notebook-loader, replace <Routes> with <Outlet/> in notebook-editor
